third_party/WebKit/LayoutTests/editing/inserting/insert-div-018.html

<!doctype html>
<body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
    assert_selection(
        [
            '<div contenteditable>',
            '<div style="height:12px"></div>',
            'bar|<div>baz</div>',
            '</div>',
        ].join(''),
        'insertParagraph',
        [
            '<div contenteditable>',
            '<div style="height:12px"></div>',
            'bar<div>|<br><div>baz</div></div>',
            '</div>',
        ].join(''),
        'Insert BR befor <div>baz</div>');
});
</script>
